Fix for pd.Series input to woe_1d
The woe_1d function wasn't running y.reset_index() in
the case where its y input was already a pd.Series. This
meant that it would return incorrect values when indexes
in X and y were noncontinuous.

Added a check for this condition to tests/test_metrics.py
as well.