Fix negative IV values in summary tables
The scikit-learn WOEEncoder uses ln(%bad / %good) instead of
ln(%good / %bad) for some reason, which flips all of the signs.
A previous fix switched the computation of WOE in the summary
table to match. However, this fix caused the IV values to be
negative in summary tables, which I found surprising.

I added an abs() call to the IV calculation to correct this
cosmetic defect.